Skip records (#79)
* new options

* new options

* process simu

* skip records and delete records

* debug info better

* debug

* default rate increase for se

* ok for failing cuts

* ok for failing cuts

* no more config

* fixed bad logic in delete_records

* avoid overwrite

* final for sr1

+++ saltax/instructions/generator.py
@@ -151,7 +151,12 @@ def instr_file_name(runid, instr, recoil, generator_name, mode, rate=1e9/SALT_TI
     filename = BASE_DIR + runid + "-" + str(recoil) + "-" + \
         generator_name + "-" + mode + "-" + str(rate) + ".csv"
     
-    pd.DataFrame(instr).to_csv(filename, index=False)
+    # if the file already exists, we don't want to overwrite it
+    if not os.path.exists(filename):
+        pd.DataFrame(instr).to_csv(filename, index=False)
+    else:
+        print("Instruction file already exists at: %s" % (filename))
+        
     print("Instruction file at: %s" % (filename))
 
     return filename
